requirements.txt
werkzeug>=3.0.1
openai>=1.0.0
flask==2.2.5
flask-cors==3.0.10
asgiref==3.7.2
python-dotenv>=1.0.0
openai==1.12.0
numpy
pandas
requests
httpx==0.24.1
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
mangum==0.17.0
supabase==1.2.0  # Supabase Python client (official)
detoxify==0.5.2
transformers==4.39.3
torch==2.2.2








